infrastructure/generative_models/requirements.txt
clr
dill
imolecule
joblib
matplotlib
nltk
numpy
pandas
Pillow
rdkit>=2018.09.1.0
scikit-learn==1.2.2
scipy
seaborn
selfies
sets
SmilesPE
spacy == 3.7.5
torchtext==0.6.0
XlsxWriter
tensorflow
protobuf
urllib3
torchdata==0.4.1
fastapi
uvicorn
lightgbm
vina
bio
huggingface_hub
catboost